+++ test/import-data/results/github-scene-template.ts
@@ -1,0 +1,36 @@
+const GITHUB_SCENE_TEMPLATE = [
+    "/ch/[CH]/config [LIB] [LIB] [LIB] [SCN]",
+    "/ch/[CH]/delay [LIB] [LIB]",
+    "/ch/[CH]/preamp [LIB] [LIB] [LIB] [LIB] [LIB]",
+    "/ch/[CH]/gate [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB]",
+    "/ch/[CH]/gate/filter [LIB] [LIB] [LIB]",
+    "/ch/[CH]/dyn [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB] [LIB]",
+    "/ch/[CH]/dyn/filter [LIB] [LIB] [LIB]",
+    "/ch/[CH]/insert [LIB] [LIB] [LIB]",
+    "/ch/[CH]/eq [LIB]",
+    "/ch/[CH]/eq/1 [LIB] [LIB] [LIB] [LIB]",
+    "/ch/[CH]/eq/2 [LIB] [LIB] [LIB] [LIB]",
+    "/ch/[CH]/eq/3 [LIB] [LIB] [LIB] [LIB]",
+    "/ch/[CH]/eq/4 [LIB] [LIB] [LIB] [LIB]",
+    "/ch/[CH]/mix [LIB] [LIB] [LIB] [LIB] [LIB] [LIB]",
+    "/ch/[CH]/mix/[IE --odd] [IEM-S] [IEM-S] [SCN] [SCN]",
+    "/ch/[CH]/mix/[IE --even] [IEM-S] [IEM-S]",
+    "/ch/[CH]/grp [LIB] [LIB]",
+    "/ch/[CH]/automix [LIB] [LIB]",
+    "/bus/[IE]/config [IEM] [IEM] [IEM]",
+    "/bus/[IE]/dyn [IEM] [IEM] [IEM] [IEM] [IEM] [IEM] [IEM] [IEM] [IEM] [IEM] [IEM] [IEM] [IEM] [IEM] [IEM]",
+    "/bus/[IE]/dyn/filter [IEM] [IEM] [IEM]",
+    "/bus/[IE]/insert [IEM] [IEM] [IEM]",
+    "/bus/[IE]/eq [IEM]",
+    "/bus/[IE]/eq/1 [IEM] [IEM] [IEM] [IEM]",
+    "/bus/[IE]/eq/2 [IEM] [IEM] [IEM] [IEM]",
+    "/bus/[IE]/eq/3 [IEM] [IEM] [IEM] [IEM]",
+    "/bus/[IE]/eq/4 [IEM] [IEM] [IEM] [IEM]",
+    "/bus/[IE]/eq/5 [IEM] [IEM] [IEM] [IEM]",
+    "/bus/[IE]/eq/6 [IEM] [IEM] [IEM] [IEM]",
+    "/headamp/[CH +31] [LIB] [LIB]"
+];
+
+export {
+    GITHUB_SCENE_TEMPLATE
+}